Condense menus in a single module
